Improve the facing API of the Token class
Previously Token saved its facing as an Integer object with a value
of -179 to 180 degrees. While there is no harm in storing it as a
negative value it is not the standard mathematical praxis. In order
to improve the usability of facing for developers new to Maptool it
is now stored as an angle 'a' so that 0 <= a < 360 degrees.

Additionally the fact that the facing is now saved in the float data
type instead of an Integer object allows for more digits of precision
while at the same time consuming less memory.

It is now also possible to set and get the facing in radians.

The old methods setFacing and getFacing have been marked as deprecated.
The way they behave have not been changed, only the internal workings
of them. From now on setFacingInDegrees and getFacingInDegrees should
be used instead.

Tokens created in Maptool 1.3 and saved without a facing gets assigned
the default float value of 0 when unserialized from XML. In other
words, they will face toward east.