Feature Enhancement - Token VBL
Tokens can now store VBL which renders independently of the normal VBL
and moves with the token. Several features were added to support Token
VBL as follows:

 - New tab on Edit Token Dialog to show Token VBL which allows:
   - Auto Generation of VBL based on Tokens
   - Clearing of Token VBL
   - Toggle for "Always Show" which will show the token over FoW when
another token can see x of it's 9 regions, with x = Visibility Tolerance
setting on token.
 - Normal VBL layer shows Token VBL as translucent yellow under normal
VBL to assist in placement
 - Added "Always Visible" menu option to Token/Stamp pop-up menu (right
clicking on token) which replaces "Block Vision" from previous release
 - Campaign export updated with new Token class fields for omission
 - Squashed a bug that prevented getExposedTokens() macro from returning
proper tokens when a Figure token or VBL Token was on the map.
  - Squashed a bug that prevented getExposedTokens() macro from
returning proper tokens if No Vision was selected at the same time FoW
was turned on.
 - New Macro functions to support new token vbl options:
   - setAlwaysVisible() & getAlwaysVisible, set/get Always Visible
option on tokens
   - getTokenVBL() & setTokenVBL(), to set/get VBL stored on tokens
   - transferVBL() directly transfers vbl from token to normal VBL if
true, otherwise it transfers from VBL to token 
 - Minor cleanup of a couple classes for imports and console debug
statements