Fix for drag/drop on OS X Java 1.8u72+
Somewhere around 1.8u60 it seems Oracle added support for additional
drag/drop data flavors.  OS X now has flavors it didn't before and
it broke.  The tweak is a better implementation overall as it allows
a data flavor to match but the import to fail, yet the next data
flavor will still be attempted (previously it would not proceed to
the next flavor on failure).