GitHub issue 37
Now catching exceptions from the heartbeat thread.  Treating them as
transient issues, but we notify the chat window of the system hosting
the server and all GMs, providing the server's IP address and port
number.  If we decide the messages happen too frequently, it's easy to
change the backoff by changing the WARNING_TIME.