Added top-level User readme file for MapTool
We have a top-level readme for developers that discusses how to create
an Eclipse environment and how to install Gradle, but lost in the
shuffle was a README for users to peruse after installing MapTool.
I'd like to eventually have options on the Help menu that show the
README file and a Release Notes file (in the beginning just a change log
for the build?).