1.4.1.13 Enhancements & Bug Fixes
 o New App Preference: Enable Initiative Panel Server Sync; Default is
off. While disabled, clients Init Panel does NOT sync with the server.
This greatly improves performance and prevents a bug where the panel can
sync and overwrite on itself causing the initiative to jump
around/backwards. (This is sort of a hack fix but it's agreed Initiative
system needs a rewrite, so until then...)
 
 o Changed herolab.getInfo function to return a Zero if no HeroLab data
exists vs a ParserException which kills the running macro.
 
 o TokenPropertiesEdit Dialog bug fixl config panel labels on right were
being cut off.
 
 o Updated gradle.properties.sample with latest variables
 
 o Updated jWrapper to 00044250826
   o Update build template to only build x64 versions