Delete temp file
This file is created tmorarily during testing, but the name  used can't be checked out on Windows so the dev branch is useless.  The testing code should be checked to ensure it deletes the file when it's done, or failing that, there should be a .gitignore in the directory that prevents it from being added to the repo.