Updated build.gradle to match Jamz' version
Some significant tweaks to turn off features that the core repo doesn't
need, like JavaFX.  Tried to use variable settings as much as possible
so that only a few lines at the top of the file need to be changed.
Commented out a bunch of stuff (sentry, box2d, etc) that could also be
enclosed in IF statements so that on/off settings could be created for
them as well.