Cleanup in preparation for macro editor code documentation
There is a problem in the way the original code handled the "Search..."
submenu:  on macOS, it was attempting to create a JMenuBar to be placed
on the JDialog, but macOS wants to do all menus at the top of the
screen.  This meant looking for the "Edit" menu installed by
MapTool.main() and adding the search options to the bottom (seemed to be
the best place for them).  The current code needs work because it looks
for "Edit" and that won't work when i18n is taken into account.