Works now, but macOS still acts a little flakey
It seems that the macOS libraries might have a race condition in them as
I sometimes I get the spinning pinwheel.  Pausing the app in Eclipse
shows that there is a thread sitting in an `await()` call, waiting for a
lock.

So this commit is to cleanup as much as possible and fix a couple bugs
in the properties file.