More scripting updates; docs added to README.md
The last step of this process is to copy the result to the resources/
directory, but I'm reluctant to do it automatically right now; it's a
manual process.  In the next phase, I'll make it automatic with the
assumption that the person running this knows how to revert changes that
are already in Git. ;)