Add support to catch an abort and a new return function
To exit a function/macro use the new "return" function.
First parameter is same as in "abort" function, so the return will only
do anything if first parameter is 0. Second parameter is an optional
return value, which will be put in macro.return and in output text of
the macro using the "return" function.

E.g.:
[return(0, "returned hello world")]
Flow will end and leave the macro continuing in the calling macro once the return function is evaluated.
To not stop all macro execution when some macro in the callstack is
calling "abort(0)" you can set a variable named "macro.catchAbort=1".
All macros in the callstack will still be aborted, except the one with
"macro.catchAbort=1" in scope.
E.g.:

[h: macro.catchAbort = 1]
[h: resultText = "defaultValue"]
[h: resultText = abort(0)]
[r: resultText]
Will not abort but instead have "defaultValue" in resultText. The abort can happen anywhere above in the callstack, so could also happen in another macro that we call.