Catches exceptions related to missing JavaFX files
MapTool currently uses JavaFX to create the splash screen.  This
prevents the JAR from running in environments where those libraries are
not available.  However, the splash screen is not the only issue -- the
JIDE libraries are so old that they also create ClassNotFound exceptions
when executed under Java 11.  This commit is the first in this series
that cleans up the splash screen requirement.  Another commit is coming
that will update the JIDE libraries as well.