Added `--html` to eliminate some entity errors
It seems that some entities, like `&bull;` for bullets, do not parse
properly without the `--html` flag.  This sort of makes sense because
those entities are not externally defined; they're part of HTML.
Doesn't explain why this program only now started complaining.  (Maybe
something to do with an app update?)