Added JIDE libraries, version 3.7.6
Documentation is sparse, but I _believe_ these will correct some Java 11
incompatibilities.  Many of the bugs reported fixed don't seem to affect
MapTool for whatever reason.  I updated the version numbers in the
Gradle build script, but the libraries themselves either need to be put
into a Maven repo listed therein, or the "lib" directory needs to be
added to the search locations (I'm not familiar with how to do that).