Refresh macro editor docs from Craig's wiki
Manually added {get,set}TerrainModifier() function names so I didn't
have to grab another wiki dump.

Couldn't get movedOverPoints.html from the downloaded HTML to process
correctly.  No idea what caused `xsltproc` to freak out, but the script
that works for everry other file caused it to fail and generate an empty
output file.  After messing with it for awhile, I just manually
extracted the proper <div>.  Bummer, because if this is necessary to do
again, the manual part will be as well.