Fixes for getTokens with JSON conditions
Use of toString() in the layer filter was causing an exception in enum.valueOf() call.  matchTokens() was failing comparisons as it was forcing the Layer name from the token to upper case where the LayerFilter was using the enum values not the names.  booleanCheck() was also producing class cast exceptions with getAsJsonPrimitive() call.  The jobj parameter could have a mix of JsonObject, JsonArray or JsonPrimitive.