Long term solution works, but...
Since the Mac mouse has inertia, it's possible to press or release the
Shift key in the middle of the inertia.  This means that MapTool goes
from spinning the facing to zooming the map just by releasing the key.
It's a little unusual.  A better UX would be to ignore all future events
based on inertia, but I'm not exactly sure how to do that.