Fixes problem with extension not being always added to Campaign filename in Export and Save
This is a fix for #1269. Since code between Export As and Save As is very similar, it has been
refactored out as much as possible.

This also fixes a problem in the Save As function: if you try to save the campaign as c1, it checks
if c1 exists, but then saves it as c1.cmpgn. Fixed.

Manual test cases:
Test Cases:

Export:
- existing file, no extension => OK
- not existing file, extension => OK
- not existing file, no extension => OK
- existing file, extension => OK

Save:
- existing file, no extension => OK
- not existing file, extension => OK
- not existing file, no extension => OK
- existing file, extension => OK

Refers to #1269