Refactors the SysInfo, gathering information asynchronously
The refactoring is actually about splitting the SysInfo class so that the code for handling the GUI
is in one class and the code for dealing with gathering information is in another class.

Also, it creates an interface (SysInfoProvider) so that the GUI depends from that.

The function getInfoFunction, which was also depending on SysInfo, now only depends on
MapToolSysInfoProvider, the concrete implementation of SysInfoProvider. This allows to test
the "debug" function there (added a couple of unit tests to ensure everything still works).

Refers to #126