Improve token image handling
Keep soft references to loaded images so they don't have to be re-regenerated unless the VM is looking for memory to garbage collect. This makes switching between zones faster.

Don't generate a gray image when dragging tokens - use an alpha composite instead. This improves select/drag performance and uses less memory.
This is a bit more tricky as playing with the Graphics2D state can affect rendering of the next token. There's more opportunity to use a simple transformation for flip/flip-iso which are still being generated.

Fixes #1801