Remove a couple no-op lines
Neither of the deleted lines of code can do anything, clearly.

The one in AnimationManager is technically a data-race in that it's
accessing a shared list across a thread boundary outside of a
synchronized block, but since it does nothing, a racy read probably
isn't a big deal.

It still seems like it's clean to delete these meaningless lines.