Implement 'getArea' for remaining Template drawings
Most Template drawings did not have a getArea implementation.

This adds a straightforward implementation to each of them.

In general, the pattern for creating these was by mimicking the
'paintArea' code in order to calculate the set of squares that would be
drawn, and then rather than paint them, merge them all into one area.

There's room for refactoring this a bit more in order to remove the
duplication between 'getArea' and 'paintArea'.
One idea would be to add a primitive to the template drawings of the
list of squares it covers, implement just that for each template, and
then have AbstractTemplate know how to implement everything else from
that.

For now, it seems pragmatic to do it with this amount of duplication and
to tackle that refactoring separately.