onChangeToken version 2
Provides additional information to a macro that is triggered using the onChangeToken event, such as update source, and any parameters that are easily representable in string form.

A user that wishes to use version 2 may include 'version+"2"' in their onChangeToken link tag, otherwise, defaults to the original implementation.

Now uses json to provide information to the listening macro.