Fix topology function error messages
First of all, the actual macro function name is used, appended with the shape being drawn. For example, instead of a reporting an error for function `drawTopology[Rectangle]`, we now report an error for `drawMBL[Rectangle]` if the error resulted from a call to `drawMBL`.

Second, the order of the arguments during i18n has been fixed so that it reads better. For example, instead of saying:
> Argument key "drawMBL[Rectangle]" to function "{x,y,w,h}" must be an integer.

We now say:
> Argument key "{x,y,w,h}" to function "drawMBL[Rectangle]" must be an integer.

This change also includes a very minor reduction in code duplication.