Make the external IP address determination be parallel.
This change introduces `getAddressAsync()` as an asynchronous alternative to `getAddress()`. `getAddressAsync()` will
attempt to read an IP address from all URLs concurrently, and will only resolve once one of them returns
successfully. The original `getAddress()` is now a wrapper around `getAddressAsync()` that waits for the IP address to
be returned, and times out after a while otherwise.