Fix AI pathfinding in the top-leftmost MBL box.
This is a workaround to address #2379.

Since PathIterator returns all holes prior to returning all shells, we reverse the order to work better with JST's
assumption that holes follow shells. Although this still does not fully fulfill JST's assumptions, it is sufficient for
JST to correctly identify all holes when combined with buffering.

This change relies on the order of PathIterator. Although the order of PathIterator appears to be consistent, I haven't
found this documented anywhere. So this change isn't really a proper fix as much as it is a workaround for JST's
behaviour as it stands with the current behaviour of PathIterator. A robust solution would require fixes in JST to
reduce or correct assumptions about PathIterator behaviour.