Avoid prematurely closing nodes in token footprint.
Nodes were being closed when MBL blocks a direct path from a footprint cell to any of the current node's
neighbors. This didn't really affect anything the current cell, since A* would end up closing it anyways. But it did
cause problems for the other cells in the footprint, and it also caused problems now that more than one node exists for
each cell.