Finished cleanup for Step 1
Step 2 is run after the web pages are updated.  Melek is going to do the
one function that needs updating, then I'll start it.  It's job is to
create a URL for each page and download the HTML text.  Steps 3 & 4 just
process that data.  Will probably need to come back to step 1 when he
creates new templates for MacroEvent and SpecialVariable, since we'll
want to grab those pages as well.