Automatically constrain the A* search space
Rather than allow A* to operate on the entire (unbounded) map space, it is now constrained to looking for solutions in
the vicinity of the MBL and terrain modifiers, as well as the start and goal cells. This relies on the fact that only a
finite number of features can influence pathfinding, and those tend to be in a small area.