Update macro editor help docs from wiki
Melek has more things to edit on the wiki to add doc for macro events
and special variables that are missing from this list.  But that
requires new wiki templates, requires updating pages, requires updates
to these scripts, and then all of that must be executed and the output
snagged for use.