Make the asset cache more robust, to avoid broken assets.
When writing to the persistent asset cache, we no longer directly write to the asset file. Instead, we write to a
temporary file (`tmp.*`) and then rename it once all data has been written to the file. This reduces the chance that an
asset file contains partial results, and makes it easy to determine whether a cache entry is valid and available for
reading.

When reading from the persistent asset cache, if the cached asset's MD5 does not match its ID, we do not load it since
it does not have valid data. Due to the write-side changes, we can also be confident that the file itself is corrupted
rather than simply being partially written. We purge the corrupted file from the cache so that it can be recached in the
future if a link to the original file still exists.

A minor change is also made to the writer so that it uses an executor instead of creating a new thread for each asset.