Remove redundant PathIterator segments to avoid self-intersections
A weakness in AWT's `PathIterator` is that it sometimes yields miniscule segments that are essentially zero-length. When
processed by JTS, these segments are interpreted as a self-intersection, which it tries to handle
intelligently. Unfortunately, doing so is difficult especially when shapes are nested.

This changes removes these near-zero-length `SEG_LINETO` segments. Doing so also the possibility that a given ring no
longer contains any `SEG_LINETO`. These degenerate rings are also removed.

In order to more easily discover broken geometry in the future. the `Geometry::buffer()` call is now done after the
logging for invalid geometry. Since `Geometry::buffer()` always creates valid geometry, the logging was not being
triggered even when tested with broken geometry.