Work around missing JavaFX symbols on linux.
On linux, the JavaFX native webkit libraries are missing the necessary symbol that provies the implementation of
`HTMLOptionsCollectionImpl::itemImpl`. As a result, we are unable to access the items returned by
`HTMLSelectElement::getOptions()`, and thus can't use it to determine which options of a `<select>` are selected.

To workaround this, we take a dependency on the `HTMLSelectElementImpl`, the JavaFX implementation of
`HTMLSelectElement`. It has a method `item()` that is not exposed in the `HTMLSelectElement` interface, but which
provides access to the associated options. Ideally we'll remove this once JavaFX is fixed, so we don't keep a dependency
on internals.