Change `VisibleAreaSegment::getPath()` to return `Path2D`.
Measurements were showing that `PointerTool::exposeFoW()` was spending about 60% of its time in
`VisibleAreaSegment::getPath()`. This methods constructs an `Area` from a path, but the callers only care about getting
the bounding box of the `Area`, and not the `Area` itself.

This change just skips the creation of the `Area` and instead returns the much cheaper `Path2D` that is created. Since
the conversion to `Area` is quite expensive, this eliminates almost all of the running time of `getPath()`, and yields
about a 2x speedup for `exposeFoW()`.