Handle HEX when updating the Light tab in Campaign Properties
"hex" is accepted as a valid shape when editing lights, and hex-shaped lights will work when attached to a token. But
when a hex light is defined, the campaign properties dialog cannot longer be opened due to a `NullPointerException` This
changes fixes that by adding `HEX` as a recognized shape when updating the lights.

There is also a slight restructuring: if we encounter an unrecognized light shape, we now deliberately throw a
`RuntimeException` rather than allowing a `NullPointerException` further along. This has the perk that we don't need to
initialize `shape` to `null`, because Java recognizes that all code paths will initialize `shape` before it is read.