Fix JTS geometry errors in calculateVisibility()
There are lots of ways in which cleared area geometry can become invalid, even though it is composed of many valid
polygons. Operations like unions are particularly susceptible to breaking as a result. To avoid issues with unions of
cleared area while also avoiding excessive buffering, we now build a `GeometryCollection` containing all of the cleared
area geometry. We then `buffer(0)` that, which has the effect of creating a single valid geometry without growing the
geometry at all.

In order to make this work robustly without having to buffer each individual cleared area, we no longer try to union
individual cleared areas together and instead only union the many cleared areas in a single `UnaryUnionOp`.  Although
this can theoretically limit the number of culled segments, performance actually improves in practice.