Rename TopologyMode to TopologyTypeSet
Campaigns that predate terrain VBL could not be loaded without error due to the refactoring in #3148. This was because
the old `Zone.topologyMode` field has the same name and type name as before the refactoring, but was structurally
distinct and so could not be correctly deserialized.

To avoid this, and to avoid potential other confusions with the previous enum `TopologyMode`, the new class has been
renamed to `TopologyTypeSet` and methods and variable names have been updated accordingly.