Fix so Edit Token Dialog does not keep state from previous tokens
Addresses #3234

There were a couple issues to resolve:
- When a token's VBL is first generated in the dialog, the `autoGenerated` is set to `true` so that the VBL will be
regenerated on various UI changes (e.g., changing the method). This flag was not being reset when a new token is
edited, meaning the new token would automatically have VBL generated and applied to it once the dialog's state is set
upon opening.
- When the Edit Token Dialog is opened, the various UI elements are initialized to match the state stored in the token
VBL panel. But since the token VBL panel wasn't reset until _after_ this initialization, the state was being set
according to the previous panel state and not the new panel state.

There is also a cleanup change to reset the calculated point counts to 0 even when the VBL is null, which is the case
for new tokens. It doesn't look like this change is currently observable, but given that there are some conditions on
the number of points being non-zero, this just reduces the possibility of future bugs due to the point count being
preserved from prior uses of the dialog.