Support temporarily disconnected WebRTCClientConnections
WebRTC is tolerant to temporarily disconnected peer connections,
which can happen if a client doesn't send the heartbeat in time,
or a network link goes down,
but this is recoverable since the client may resume responding
and other routes may be active.

For this reason there is a separate RTCPeerConnectionState of
FAILED, which is distinct from DISCONNECTED
because a DISCONNECTED connection may become re-CONNECTED.

To handle potential disconnections, the sendThread is put to sleep if
there are no more messages or the peer connection is DISCONNECTED,
and is woken up on either sendMessage or connection state change
if there are both messages and it is CONNECTED.

RTCDataChannel.send may throw an Exception if its buffer is full,
which may be more likely to happen if the connection is disconnected,
but is not expected to throw while disconnected.