Restrict A* to exposed areas when FoW is enabled.
This change prevents pathfinding from showing paths through hard FoW, so as to avoid spoiling details unknown to the
player. Analogous to the existing MBL check, the new FoW check will require that the line from the center of one cell to
the next be completely exposed in order to allow that move.

The area that is considered exposed relies on the server settings (**Use Individual FoW**) as well as the effective
player role. I.e., if the GM has enabled **Show as Player**, pathfinding will be restricted from passing through hard
FoW. Otherwise, the GM is never restricted by FoW, and pathfinding will only account for MBL.

Fix names to refer to exposed area as exposed area and not as hard FoW