Calculate lights when map vision is set to Day or Off
In addition, the visible area will be treated as a source of light with a lumens value of 1. This means any darkness can
override it while still enabling PCs with the necessary light to see into the darkness.

Also refactor some of the light area calculations so that sorting by lumens and handling negative lumens is a
responsibility only for `getVisibleArea()` as that is the place that actually cares about the sort order. Sorting is now
handled by comparator rather than carefully inserting certain keys for positive lumens and other keys for negative
lumens. This means we can easily add to the map without first having to ensure the lumens meets the sorting
requirements.