Render all lights with BlendingComposite
Lights and auras are now both rendered using the `BlendingComposite` class, so that overlapping lights and auras get
brighter rather than darker. The rendering code for lights and auras has also been refactored so that they share the
same implementation.

Lights are no longer grouped by paint. This saves a bit of work preparing for the render, and leaves the door open for
deliberate light ordering if we ever change to an order-dependent blending operator. Lights are still rendered to their
own overlay prior to being drawn to the map using alpha compositing.

Auras are no longer merged by paint. This means that overlapping auras will visibly blend instead of coalescing into a
single area. Auras are also now rendered to an image as for lights, and as a result can blend with each other prior to
being drawn to the map using alpha compositing.

Since lights and auras are rendered to their own overlays, they will not blending with each other as lights, but will
instead blend as transparent overlays are they are rendered to the map.