Do not assume a server connection exists when retrieving images
This addresses #3425. It can happen that early in the initialization process, assets are requested to populate the UI
even before any personal server has been started or `ServerCommand` created. In this case, we cannot rely on a server
call to resolve and image and call pending listeners. To work around this problem, we fallback to using a broken image
if the server is not available.