Explicitly initialize some more Token fields to avoid nulls
The `Token` class expects to have a non-null shape, type and layer. Up until now, the fields could be `null`, although
the accessors for these fields would detect this and return a real value (`readResolve()` does similar). This change
makes it so that the Token class' initializer also sets these fields to their default values, so that all constructors
will respect the defaults as well. This also means there is no more worry about these fields being null, which means the
accessors can be simplified some and `toDto()` does not also have to handle `null`s.

In a similar vein, we can infer from the `readResolve()` logic and usage that `name` and `exposedAreaGUID` do not need
to support a `null` value either, and so this change also updates these fields to be initialized to real values.