Add UVTT topology and lights to zone before adding to campaign
Previously VBL/MBL and light tokens would no be synced to clients since the new `Zone` was modified directly rather than
being updated via server commands. This change avoids the issue by performing all the modifications to the zone first,
then adding the zone to the campaign.

One other minor improvement here is that the UVTT map's topology and lights are no longer processed if the user cancels
the import from the dialog..