Update Edit Token dialog to support all token topologies
The Edit Token dialog now allows setting any and all topologies (VBL/MBL) based on that state of the new toggle
buttons. These new buttons allow independently selecting topology types, just like the toggles in the main toolbar. The
initial selection of the buttons is based on which topologies the token already has present.

A couple of minor fixes:
1. Token topology is rendered on the map in the same order as the map topology.
2. Tokens now only have a single colour sensitivity. It wasn't reasonable to support distinct sensitivities for each
   one, nor does that seem to provide any benefit.

The term "VBL" is also replaced with the more general "topology" wherever appropriate in the `EditTokenDialog`.