Add a new overlay for darkness to improve its rendering.
There are now two different cases for darkness:
1. When rendering in a GM view, darkness will blend together in a way that it gets darker when multiple sources are
   overlapping. The darkness is rendered according to the darkness overlay opacity and will be rendered on top of lights
   so as to visually dim them. In this case, neither darkness, light, nor auras will be clipped to the visible
   region.
2. When rendering in a player view, darkness will always render as opaque black so as to hide the area from the
   players. Light and aura overlays are still clipped to the visible area, but darkness is now clipped to the
   complement. This makes darkness "retreat" in the presence of powerful light or sight.

This change also fixes a bug where darkness without a color was being treated as a bright light and was not being
rendered. This has been changed so that darkness is never considered bright and will therefore always be
rendered. Darkness without a colour is treated as though its colour is pure black.