Improve performance of light rendering
A pool of device-compatible `BufferedImages` is maintained by `BufferedImagePool` that can be used whenever an image
buffer is needed that is that size of the `ZoneRenderer`. This change puts it to use for rendering the entire zone to a
buffer as well as for rendering the light/darkness/aura overlays.

More timings have been added around lighting rendering, and the timings have been renamed so that they are more
descriptive and not reliant on number sequences.

As an optimization, in the case that no light sources are present on a light/darkness/aura overlay, that overlay will
not be rendered.