Snap AreaMeta vertices to a precision grid
This grid is separate from MapTool's grid and is provided by `GeometryUtil.getPrecisionModel()`. By snapping to this
grid, we avoid needing to check points for "closeness" rather than equality while still avoiding problems with points
that would otherwise lie close to one another without technically being the same point.

The hardcoded precision is high enough that any intentional detail is going to be captured. But it is low enough that
any points that would cause issues for the vision blocking algorithm are elided.