Add a converter to make CellPoint from pre-1.9.0 AStarCellPoint XML
There are two cases in which an `AStarCellPoint` can end up having being saved:
- In campaigns created before 1.9.0, `AStarCellPoint` could be freely added to `Path` as it was a subclass of
  `CellPoint`/`AbstractPoint`.
- In campaigns loaded in 1.9.0 or later, if it already had an `AStarCellPoint` in a `Path` it would be deserialized to a
  mostly meaningless object (`null` position in particular). If that campaign is then saved again, the serialized
  `AStarCellPoint` will not have any `CellPoint` fields anymore.

In the former case, the converter builds a `CellPoint` so that the position information is not lost. In the latter
case. the information is already lost so we don't build anything. To account for this, `Path` will clear itself if it
encounters any null points, and `Token` will `null` out any empty `Path`.