Calculate visibility separately for each topology type.
This avoids the pathological case in #3626 and others that can arise from unioning sets of line segments. Many of these
cases are handled (noded and dissolved) by JTS, but there are some that cause it trouble depending on how precise the
set up is (e.g., two diagonal line segments where one is a proper subset of the other). These cases don't arise within a
single topology since they are flat areas, so we should be safe to just handle each one separately.