Remove additive blending for lights, auras and darkness
Lights are still composited into an image that is then rendered on top of the zone. The opacity is determined during
light compositing instead of when rendering the overlay image on to the zone. The effect is very similar to 1.11.5,
except that blending is performed on lights of the same colour.