Replace a bad use of Transparency constants
The `fogBuffer` was being created with either the `Transparency.TRANSLUCENT` or `Transparency.BITMASK` flags, which are
just `int`s (`3` and `2` respectively). But the constructor of `BufferedImage` actually interprets these `int`s as one
of the `BufferedImage.TYPE_*` flags (`3 == TYPE_INT_ARGB` and `2 == TYPE_INT_ARGB_PRE`). So in reality all this did was
make the fog buffer either be a premultiplied ARGB image or a non-premultiplied ARGB image depending on whether a GM or
player view active.

This change just make it so that the fog buffer is always created as `TYPE_INT_ARGB_PRE` regardless view.