Replace EventDispatcher with Guava's EventBus
We're already using `EventBus` for some things, but there were others on `EventDispatcher`. Now only `EventBus`
remains with `EventDispatcher`, `AppEvent` and `AppEventListener` being no more.

The mapping from the previous event tags to the current event types is:
- `MapTool.ZoneEvent.Activated`: `ZoneActivated`
- `MapTool.ZoneEvent.Deactivated`: `ZoneDeactivated`
- `MapTool.ZoneEvent.Added`: `ZoneAdded`
- `MapTool.ZoneEvent.Removed`: `ZoneRemoved`
- `MapTool.PreferencesEvent.Changed`: `PreferencesChanged`

The `Deactivated` was previously unused since the `Activated` event could contain an old and new value in the
`AppEvent`. But it's cleaner in the new system for subscribers to handle deactivation and activation separately and
allows the `ZoneActivated` and `ZoneDeactivated` types to be focused on a single zone. So we now conditionally post a
`ZoneDeactivated` event prior to `ZoneActivated` when necessary.