Replace the uses of ObservableList with an EventBus solution
Instead of directly listening to `ObservableList`, we instead send out events for subscribers to consume.

- ChatMessageAdded
- PlayerConnected
- PlayerDisconnected
- ServerStopped

This also means `MapTool` no longer needs to maintain a list of messages since the panel can directly listen for and
maintain the list.

`ObservableList` is no longer necessary and has been replaced with `List` where used.