Replace model events with EventBus
With this change, subscribers no longer need to subscribe to a particular model object but can generically listen for
event types that come from anywhere. In particular, this frees several components from having to ensure that their
current zone is the one they are actually listening to. For any components that relied on a particular zone, these
include a check that the event's zone matches the one in the component.

For the most part, this change does not change behaviour beyond handling events in the new way. There are a few
exceptions where components unconditionally responded to every model event whether or not it was actually
applicable. These components are now more constrained in what they respond to:
- `UndoPerZone` no longer subscribes to anything as it did nothing but log all of a zone's events.
- `DrawPanelTreeModel` only subscribes to the `DrawableChanged` event.
- `TokenPanelTreeModel` only subscribes to the various `Token*` events.
- `ZoneRenderer` only subscribes to the events that it already handled explicitly. But it no longer calls
  `updateTokenTree()` and `repaintDebouncer.dispatch()` for all the other possible events.

All the events that were fired this way were in the `Zone.Event` enum (which no longer exists). Here is the mapping from
old to new:
- `Zone.Event.TOKEN_ADDED`: `TokenAdded`
- `Zone.Event.TOKEN_REMOVED`: `TokenRemoved`
- `Zone.Event.TOKEN_CHANGED`: `TokenChanged`
- `Zone.Event.GRID_CHANGED`: `GridChanged`
- `Zone.Event.DRAWABLE_ADDED`: `DrawableAdded`
- `Zone.Event.DRAWABLE_REMOVED`: `DrawableRemoved`
- `Zone.Event.FOG_CHANGED`: `FogChanged`
- `Zone.Event.LABEL_ADDED`: `LabelAdded`
- `Zone.Event.LABEL_REMOVED`: `LabelRemoved`
- `Zone.Event.LABEL_CHANGED`: `LabelChanged`
- `Zone.Event.TOPOLOGY_CHANGED`: `TopologyChanged`
- `Zone.Event.INITIATIVE_LIST_CHANGED`: `InitiativeListChanged`
- `Zone.Event.BOARD_CHANGED`: `BoardChanged`
- `Zone.Event.TOKEN_EDITED`: `TokenEdited`
- `Zone.Event.TOKEN_MACRO_CHANGED`: `TokenMacroChanged`
- `Zone.Event.TOKEN_PANEL_CHANGED`: `TokenPanelChanged`

There were some existing events for EventBus that have been replaced with easier-to-use alternatives that are consistent
with the other events:
- `TokensAddedEvent`: `TokensAdded`
- `TokensRemovedEvent`: `TokensRemoved`
- `TokensChangedEvent`: `TokensChanged`

In conjunction with that, the `TokenEventBusBridge` is no more. There was some important functionality in there to post
token events in response to corresponding zone events, but that is now handled by posting the token events at the some
point that the zone events are posted. This functionality was also added to `ServeMessageHandler` for good measure.

Finally, since `BaseModel` only offered functionality for model events, it has been removed and existing uses (`Zone`,
`Token`) no longer have a base class. Of course, `ModelChangeListener` is also gone.