Simple case reduction by merging identical code paths
After removing `tempArea` and recognizing that the `combinedView.isEmpty()` case didn't add anything, it was clear that
three of the four cases were really the same thing and could be combined.

The fourth case really is distinct from the others as it does not render according to the current player view. In
practice, it only renders any previously exposed area for owned tokens that no longer have sight.